requirements.txt
asgiref==3.9.2
Django==5.2.6
django-environ==0.12.0
djangorestframework==3.15.2
numpy==2.3.3
pgvector==0.4.1
psycopg==3.2.10
psycopg-binary==3.2.10
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2